pyproject.toml
[project]
name = "plana-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aioboto3>=15.0.0",
    "asyncpg>=0.30.0",
    "emoji>=2.14.1",
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pydantic>=2.11.7",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "redis>=6.2.0",
    "snowflake-id>=1.0.2",
    "sqlalchemy[asyncio]>=2.0.41",
    "uvicorn>=0.35.0",
]


[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "ruff>=0.12.3",
]


[tool.semantic_release]
tag_format = "v{version}"
version_toml = ["pyproject.toml:project.version"]
version_variables = ["plana/_version.py:__version__"]
build_command = "python -m pip install build && python -m build"
commit_parser = "conventional"
commit_message = "chore(release): bump version to v{version} [skip ci]"
commit_author = "github-actions[bot] <github-actions[bot]@users.noreply.github.com>"